removed graph cut winner option, changed saliency input back to clampeddiffmean and minor variable renaming
Former-commit-id: 7baeaa1dd420f56954c684bfca2ddd6f69836847

+++ aved-mbarivision/src/main/cpp/src/DetectionAndTracking/MbariVisualEvent.C
@@ -1105,7 +1105,7 @@ namespace MbariVisualEvent {
 
                     // round down the radius in case near the edges
                     if (ebox.width() > 0 && ebox.height() > 0) {
-                        int radius = sqrt(pow((double) ebox.width(), 2.0) + pow((double) ebox.height(), 2.0))/2;
+                        int radius = (int) sqrt(pow(ebox.width(), 2.0) + pow(ebox.height(), 2.0))/2;
                         drawCircle(img, ctr, radius, colorPred);
                         if (showEventLabels) { 
                             Point2D<int> numLoc = getLabelPosition(img.getDims(), ebox, textImg.getDims());
